added "listgroupschanged" RPC method
this method takes a optional string parameter and returns one of the following
objects:
    {"groups": groups, "hash": hash}
        or
    {"hash": hash}
where hash is a string identifing groups and groups is identical to the data
returned by a call to "listgroups". The returned hash can be used in later
calls to "listgroupschanged". groups is absent in the returned object if the
hash given as a parameter is equal to the new one generated for groups because
the data that would be returned is exactly the same as it was returned in a
previous call.
An absent parameter in the call is treated as a string of length zero.