Add libcurl support for HTTP(S) downloads
Motivation
==========

This introduces support for using libcurl to handle HTTP(S) downloads
instead of the built-in TLS and HTTP client. The main driver for this is
poor handling for some HTTP(S) servers from the internal code. A
specific example is api.nzbgeek.info, which at the time of this patch
does not work with nzbget over HTTPS.

The web server at api.nzbgeek.info reacts to a SNI (server name indication)
sent by the HTTPS client with a "The server name sent was not
recognized" TLS warning message during the handshake. This is a warning
message, and it indicates that the web server is not configured
properly, but it is not necessarily fatal. The web server then proceeds
to present a valid certificate for the api.nzbgeek.info host name. All
browsers and other download tools (like curl and wget) do not treat this
warning as fatal, and continue the HTTPS handshake.

At least when compiled with GnuTLS as the SSL engine nzbget treats the
warning as fatal, and aborts the connect, making it impossible to
download files via HTTPS from api.nzbgeek.info.

Patching the TLS code path in nzbget to not treat this error as fatal is
not overly difficult, but once that is done the download does still not
succeed. This time the content is downloaded, but the interaction
between the HTTP client and the SSL layer for tearing down the
connection is slightly flawed, leading to another SSL alert ("TLS
session not properly terminate"), which is also treated as fatal.

At this point it seemed prudent to investigate offloading the whole
HTTP(S) handling to a dedicated and tested library.

Changes
=======

libcurl support is added as optional into the build process. By default
it is not enabled, leaving an unchanged version of the normal TLS/HTTP
client in place. It needs to be explicitly enabled by passing
--with-libcurl to configure.

A file is added from the libcurl distribution (libcurl.m4) to handle
detection of the library and the needed parameters. libcurl.m4 is
licensed under MIT. This should not be a problem.

The majority of other changes are to the
daemon/connect/WebDownloader.cpp file. This contains the majority of
code that interacts with libcurl proper. An effort has been made to
reuse as much of the existing infrastructure, and provide a clean
separation between "classic" and "libcurl" code.

Small changes ere made to daemon/feed/FeedCoordinator.cpp and
daemon/queue/UrlCoordinator.cpp to deal with subclasses to the
WebDownloader class, which might also initiate HTTPS connections. As
libcurl is a C library, callbacks into C++ code require some shims.

While the WebDownloader class code is tested somewhat, the
UrlCoordinator and FeedCoordinator code is not, for lack of knowledge as
to how that code path is triggered.